Rework for better accessibility
WEB-1660. We have a list of links in three columns, and with a narrow
width we only want to show the top item in each column. This was
implemented by breaking it into three lists, one for each column, and
giving the first item in each list a special class that kept it from
being hidden. There is an implied hierarchy that is not clear in this
implementation and cannot be made clear to screen readers with this
implementation.

This commit reconfigures these links as one list containing the three
top items. Each of these items has a child list containign the rest of
te items. A screen reader will announce this as a list with three items,
corresponding to the the three visual columns. It will then read the
first item and announce the child list with X items. In this arrangement
the verbal and visual hierarchies match.

When the sub-lists are not displayed visually, they hidden in such a way
that a screen reader will still read them.